requirements.txt
pandas
Django
django-cors-headers
djangorestframework
